requirements.txt
tokenizers
torchtext
lightning
datasets
tensorboard
torchmetrics